make hubcontroller an inner class of abstractirisapplication
